+++ app/views/admin/fixity/index.html.erb
