+++ app/views/sufia/admin/_sidebar.html.erb
@@ -32,8 +32,8 @@
       </li>
     </ul>
   </li>
-  <li class="<%= 'active' if current_page?'/admin/fixity'%>">
-    <%= link_to 'fixity' do %>
+  <li class="<%= 'active' if current_page? admin_fixity_index_path %>">
+    <%= link_to admin_fixity_index_path do %>
       <span class="fa fa-check-square-o"></span> <%= t('Fixity checks') %>
     <% end %>
   </li>
